make shared parent package for write and ETL packages so they can inherit common things like database and model
