+++ mETL/hello_world.py
@@ -1,14 +1,14 @@
 import psycopg2 as pg
 from config import read_params
 
-from write.model import Model
-from write.database import Database
-from write.column_data_type import IntegerColumn, TextColumn
+from write.write_model import WriteModel
+from write.write_database import WriteDatabase
+from column_data_type import IntegerColumn, TextColumn
 from write.queue import Queue
 
 
 def level_zero():
-    params = read_params()
+    params = read_params(filename='database.ini', section='localhost')
     conn = pg.connect(host=params['host'], database=params['database'], user=params['user'])
     cur = conn.cursor()
     cur.execute("create table test_table(num_col INT, text_col TEXT)")
@@ -19,16 +19,16 @@ def level_zero():
 
 
 def level_one():
-    class TestTable(Model):
+    class TestTable(WriteModel):
         def __init__(self):
             metl_queue = Queue('mETL.fifo')
-            Model.__init__(self, database=Database(queue=metl_queue, database='mETL'), table_name='test_table')
+            WriteModel.__init__(self, database=WriteDatabase(queue=metl_queue, database='mETL'), table_name='test_table')
 
         num_col = IntegerColumn()
         text_col = TextColumn()
 
     test_table = TestTable()
-    params = read_params()
+    params = read_params(filename='database.ini', section='localhost')
     conn = pg.connect(host=params['host'], database=params['database'], user=params['user'])
     cur = conn.cursor()
     cur.execute(test_table.create_sql())
@@ -39,10 +39,10 @@ def level_one():
 
 
 def level_two():
-    class TestTable(Model):
+    class TestTable(WriteModel):
         def __init__(self):
             metl_queue = Queue('mETL.fifo')
-            Model.__init__(self, database=Database(queue=metl_queue, database='metl'), table_name='test_table')
+            WriteModel.__init__(self, database=WriteDatabase(queue=metl_queue, database='metl'), table_name='test_table')
 
         num_col = IntegerColumn()
         text_col = TextColumn()
